Added payment screen and more
Added Payment Screen
When ordering a ticket, select time and date first, it goes to ticket
screen, then select seat screen, then payment screen and as last the
final screen that goes back to home.
Added the slider test from Koen into ticket screen
Added that you have to select a seat in order to go to the payment
screen.